refactor: consolidate DataContextShell -> DataContext
