+++ packages/data-context/src/DataContext.ts
@@ -1,5 +1,8 @@
 import type { LaunchArgs, OpenProjectLaunchOptions, PlatformName } from '@packages/types'
-import type { AppApiShape, ProjectApiShape } from './actions'
+import fsExtra from 'fs-extra'
+import path from 'path'
+
+import { AppApiShape, DataEmitterActions, ProjectApiShape } from './actions'
 import type { NexusGenAbstractTypeMembers } from '@packages/graphql/src/gen/nxs.gen'
 import type { AuthApiShape } from './actions/AuthActions'
 import debugLib from 'debug'
@@ -15,14 +18,19 @@ import {
   StorybookDataSource,
   CloudDataSource,
   ConfigDataSource,
+  EnvDataSource,
+  GraphQLDataSource,
+  HtmlDataSource,
+  UtilDataSource,
 } from './sources/'
 import { cached } from './util/cached'
-import { DataContextShell, DataContextShellConfig } from './DataContextShell'
 import type { GraphQLSchema } from 'graphql'
+import type { Server } from 'http'
+import type { AddressInfo } from 'net'
 
 const IS_DEV_ENV = process.env.CYPRESS_INTERNAL_ENV !== 'production'
 
-export interface DataContextConfig extends DataContextShellConfig {
+export interface DataContextConfig {
   schema: GraphQLSchema
   os: PlatformName
   launchArgs: LaunchArgs
@@ -39,11 +47,13 @@ export interface DataContextConfig extends DataContextShellConfig {
   projectApi: ProjectApiShape
 }
 
-export class DataContext extends DataContextShell {
+export class DataContext {
   private _coreData: CoreDataShape
+  private _gqlServer?: Server
+  private _appServerPort: number | undefined
+  private _gqlServerPort: number | undefined
 
   constructor (private _config: DataContextConfig) {
-    super(_config)
     this._coreData = _config.coreData ?? makeCoreData()
   }
 
@@ -170,8 +180,63 @@ export class DataContext extends DataContextShell {
     return new CloudDataSource(this)
   }
 
+  @cached
+  get env () {
+    return new EnvDataSource(this)
+  }
+
+  @cached
+  get emitter () {
+    return new DataEmitterActions(this)
+  }
+
+  graphqlClient () {
+    return new GraphQLDataSource(this, this._config.schema)
+  }
+
+  @cached
+  get html () {
+    return new HtmlDataSource(this)
+  }
+
+  @cached
+  get util () {
+    return new UtilDataSource(this)
+  }
+
   get projectsList () {
     return this.coreData.app.projects
+  }
+
+  // Servers
+
+  setAppServerPort (port: number | undefined) {
+    this._appServerPort = port
+  }
+
+  setGqlServer (srv: Server) {
+    this._gqlServer = srv
+    this._gqlServerPort = (srv.address() as AddressInfo).port
+  }
+
+  get appServerPort () {
+    return this._appServerPort
+  }
+
+  get gqlServerPort () {
+    return this._gqlServerPort
+  }
+
+  // Utilities
+
+  @cached
+  get fs () {
+    return fsExtra
+  }
+
+  @cached
+  get path () {
+    return path
   }
 
   get _apis () {
@@ -219,7 +284,7 @@ export class DataContext extends DataContextShell {
   }
 
   async destroy () {
-    super.destroy()
+    this._gqlServer?.close()
 
     return Promise.all([
       this.util.disposeLoaders(),
