+++ packages/data-context/src/sources/EnvDataSource.ts
@@ -1,10 +1,10 @@
-import type { DataContextShell } from '../DataContextShell'
+import type { DataContext } from '../DataContext'
 
 /**
  * Centralizes all of the "env"
  */
 export class EnvDataSource {
-  constructor (private ctx: DataContextShell) {}
+  constructor (private ctx: DataContext) {}
 
   get CYPRESS_INTERNAL_VITE_APP_PORT () {
     return process.env.CYPRESS_INTERNAL_VITE_APP_PORT
