+++ packages/data-context/src/sources/GraphQLDataSource.ts
@@ -2,7 +2,7 @@ import { createClient, Client, dedupExchange, ssrExchange } from '@urql/core'
 import { cacheExchange } from '@urql/exchange-graphcache'
 import { executeExchange } from '@urql/exchange-execute'
 import type { GraphQLSchema } from 'graphql'
-import type { DataContextShell } from '../DataContextShell'
+import type { DataContext } from '../DataContext'
 import type * as allOperations from '../gen/all-operations.gen'
 
 type AllQueries<T> = {
@@ -13,7 +13,7 @@ export class GraphQLDataSource {
   private _urqlClient: Client
   private _ssr: ReturnType<typeof ssrExchange>
 
-  constructor (private ctx: DataContextShell, private schema: GraphQLSchema) {
+  constructor (private ctx: DataContext, private schema: GraphQLSchema) {
     this._ssr = ssrExchange({ isClient: false })
     this._urqlClient = this.makeClient()
   }
