+++ packages/server/lib/server-e2e.ts
@@ -16,7 +16,8 @@ import * as ensureUrl from './util/ensure-url'
 import headersUtil from './util/headers'
 import statusCode from './util/status_code'
 import type { Cfg } from './project-base'
-import { DataContextShell } from '@packages/data-context/src/DataContextShell'
+import { makeLegacyDataContext } from './makeDataContext'
+import type { DataContext } from '@packages/data-context'
 
 type WarningErr = Record<string, any>
 
@@ -45,7 +46,7 @@ const isResponseHtml = function (contentType, responseBuffer) {
 export class ServerE2E extends ServerBase<SocketE2E> {
   private _urlResolver: Bluebird<Record<string, any>> | null
 
-  constructor (ctx: DataContextShell = new DataContextShell()) {
+  constructor (ctx: DataContext = makeLegacyDataContext()) {
     super(ctx)
 
     this._urlResolver = null
